refactor tests to comply with classes that now correctly hold methods that were in wrong class previously
